% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{allVSall_pid}
\alias{allVSall_pid}
\title{Calculate all-vs-all protein percent identities}
\usage{
allVSall_pid(protlist)
}
\arguments{
\item{protlist}{List of protein sequences}
}
\value{
Dataframe with percent identities
}
\description{
Calculate all-vs-all protein percent identities
}
