% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{calculate_pid}
\alias{calculate_pid}
\title{Calculate pairwise percent identity between two sequences}
\usage{
calculate_pid(
  seq1,
  seq2,
  substitutionMatrix = "BLOSUM62",
  gapOpening = 10,
  gapExtension = 0.5,
  type = "global",
  ...
)
}
\arguments{
\item{seq1}{First sequence}

\item{seq2}{Second sequence}

\item{substitutionMatrix}{Substitution matrix to use for alignment}

\item{gapOpening}{Gap opening penalty}

\item{gapExtension}{Gap extension penalty}

\item{type}{Type of alignment. One of "global", "local", "overlap"}

\item{...}{Additional arguments passed to pwalign::pairwiseAlignment}
}
\value{
Pairwise percent identity
}
\description{
Calculate pairwise percent identity between two sequences
}
