% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{ggsave_portrait}
\alias{ggsave_portrait}
\title{ggsave with A4 size (portrait)}
\usage{
ggsave_portrait(filename, plot = last_plot(), ...)
}
\arguments{
\item{filename}{character. The name of the file to save the plot to.}

\item{plot}{ggplot object}

\item{...}{Other arguments passed to ggsave()}
}
\description{
This function saves the plot with A4 size in portrait mode. rksfun
}
