% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{\%NOTIN\%}
\alias{\%NOTIN\%}
\title{Negate the \%in\% operator}
\usage{
x \%NOTIN\% y
}
\arguments{
\item{x}{A vector of values to check if they exist in y}

\item{y}{A vector of values to check if x values exist in}
}
\description{
This function negates the \%in\% operator.
I really don't know why this is not a built-in operator in R
rksfun
}
