% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{hasOneOfMultiplePatterns}
\alias{hasOneOfMultiplePatterns}
\title{Function to get a logical vector if a text contains one of multiple sting patterns}
\usage{
hasOneOfMultiplePatterns(patterns, text)
}
\arguments{
\item{patterns}{A vector of strings to detect in the text}

\item{text}{A vector of strings to detect the patterns}
}
\value{
A logical vector of the same length as the text vector
}
\description{
Function to get a logical vector if a text contains one of multiple sting patterns
}
\examples{
# define a vector of patterns
patterns <- c("Merc", "Hornet")
# create a DF
df <- datasets::mtcars \%>\% tibble::rownames_to_column(var = "car")
# filter the DF to keep only the rows that do not contain one of the patterns
df_with_pattern <- df \%>\% dplyr::filter(rowHasPattern(patterns,df$car))
# filter the DF to keep only the rows that do not contain one of the patterns
df_without_pattern <- df \%>\% dplyr::filter(!rowHasPattern(patterns,df$car))
}
