% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_gen}
\alias{palette_gen}
\title{rksfun palette generator for discrete palettes}
\usage{
palette_gen(palette = "rktbns_post", direction = 1)
}
\arguments{
\item{palette}{character name of the palette}

\item{direction}{integer 1 for normal order, -1 for reversed order}
}
\value{
function that returns a character vector of colors
}
\description{
Function to generate a function that returns a palette of a given length
}
