% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{read_annotation}
\alias{read_annotation}
\title{Read the result file from my annotation pipeline}
\usage{
read_annotation(file, threads = 12, ...)
}
\arguments{
\item{file}{character. The name of the file to read}

\item{threads}{integer. Number of threads to use}

\item{...}{Other arguments passed to read_tsv}
}
\description{
This function reads the result file from my annotation pipeline
}
