% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rks_palette}
\alias{rks_palette}
\title{rksfun palette}
\usage{
rks_palette(palette = "rktbns_post", ...)
}
\arguments{
\item{palette}{character name of the palette}
}
\value{
character vector of colors
}
\description{
Function to return hex codes for the included palettes
}
