% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rksfun-package.R
\docType{package}
\name{rksfun-package}
\alias{rksfun}
\alias{rksfun-package}
\title{rksfun: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Richard Stöckl \email{richard.stoeckl@ur.de} (\href{https://orcid.org/0000-0002-0451-0652}{ORCID})

}
\keyword{internal}
