% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rksfun.pals}
\alias{rksfun.pals}
\title{rksfunPalettes for plotting}
\usage{
rksfun.pals(
  name,
  n,
  type = c("discrete", "continuous"),
  direction = c(1, -1),
  override.order = FALSE
)
}
\arguments{
\item{name}{Name of Palette. Choices are:
\code{rktbns_post}, and \code{tf2}}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{override.order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}
}
\value{
A vector of colors.
}
\description{
rksfunPalettes for plotting
}
