% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{rksfunPalettes}
\alias{rksfunPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 2.
}
\usage{
rksfunPalettes
}
\description{
Use names(rksfunPalettes) to return all possible palette names. Current choices are:
\code{rktbns_post}, and \code{tf2}
Use \code{\link{rksfun.pals}} to construct palettes.
}
\keyword{datasets}
