% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_rksfun_c}
\alias{scale_color_rksfun_c}
\title{rksfun continuous colour scale for ggplot2}
\usage{
scale_color_rksfun_c(palette = "rktbns_post", direction = 1, ...)
}
\arguments{
\item{palette}{character name of the palette}

\item{direction}{integer 1 for normal order, -1 for reversed order}
}
\value{
ggplot2 scale function
}
\description{
Function to create a continuous ggplot2 colour scale function for the rksfun palettes
}
