% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_rksfun}
\alias{scale_colour_rksfun}
\title{rksfun color scale for ggplot2}
\usage{
scale_colour_rksfun(palette = "rktbns_post", direction = 1, ...)
}
\arguments{
\item{palette}{character name of the palette}

\item{direction}{integer 1 for normal order, -1 for reversed order}
}
\value{
ggplot2 scale function
}
\description{
Function to create a ggplot2 color scale function for the rksfun palettes
}
