% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{\%NOTIN\%}
\alias{\%NOTIN\%}
\title{Negate the \%in\% operator}
\usage{
... \%NOTIN\% NA
}
\arguments{
\item{...}{vector}
}
\description{
This function negates the \%in\% operator.
I really don't know why this is not a built-in operator in R
rksfun
}
